requirements.txt
asgiref==3.8.1
coverage==7.8.0
Django==5.2
django-ckeditor-5==0.2.17
django-environ==0.12.0
django-js-asset==3.1.2
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flake8==7.2.0
mccabe==0.7.0
pillow==11.1.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
